fix : addToCart분리, calculateRemainingStock 분리
