+++ src/basic/models/cart/useCart.ts
@@ -1,0 +1,66 @@
+import { useCallback, useState } from "react";
+import { CartItem, Product } from "../../../types";
+import { ProductWithUI } from "../products/types";
+import { NotificationType } from "../../hooks/useNotifications";
+import { calculateRemainingStock } from "../../utils/calculateRemainingStock";
+
+export const useCart = (
+  addNotification: (message: string, type: NotificationType) => void
+) => {
+  const [cart, setCart] = useState<CartItem[]>(() => {
+    const saved = localStorage.getItem("cart");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return [];
+      }
+    }
+    return [];
+  });
+
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = calculateRemainingStock(product, cart);
+      if (remainingStock <= 0) {
+        addNotification("재고가 부족합니다!", "error");
+        return;
+      }
+
+      setCart((prevCart) => {
+        const existingItem = prevCart.find(
+          (item) => item.product.id === product.id
+        );
+
+        if (existingItem) {
+          const newQuantity = existingItem.quantity + 1;
+
+          if (newQuantity > product.stock) {
+            addNotification(
+              `재고는 ${product.stock}개까지만 있습니다.`,
+              "error"
+            );
+            return prevCart;
+          }
+
+          return prevCart.map((item) =>
+            item.product.id === product.id
+              ? { ...item, quantity: newQuantity }
+              : item
+          );
+        }
+
+        return [...prevCart, { product, quantity: 1 }];
+      });
+
+      addNotification("장바구니에 담았습니다", "success");
+    },
+    [cart, addNotification, calculateRemainingStock]
+  );
+
+  return {
+    cart,
+    setCart,
+    addToCart,
+  };
+};
