+++ src/basic/utils/calculateRemainingStock.ts
@@ -1,0 +1,9 @@
+import { CartItem, Product } from "../../types";
+
+export const calculateRemainingStock = (
+  product: Product,
+  cart: CartItem[]
+): number => {
+  const cartItem = cart.find((item) => item.product.id === product.id);
+  return product.stock - (cartItem?.quantity || 0);
+};
